requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-decouple==2.1
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
Markdown==3.3.7
pip==22.1.2
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.28.1
setuptools==59.6.0
sqlparse==0.4.2
urllib3==1.26.10
